requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.7
psycopg2-binary==2.9.7
reportlab==4.0.7
python-dotenv==1.0.0
WTForms==3.0.1
email-validator==2.0.0
Jinja2==3.1.2
